New resource for integrating firebase with React
This series helped me more than any other in using firebase with React. Also was made fairly recently so almost everything in the videos are identical to the current versions of React and firebase. The guy also has a few videos on how to test React apps that were really straight forward